fix(timetables): fix null stoptime in trip sorter
The first stopTime for a trip was being used to sort the trips returned from the server. When that
stopTime was null, the sorter was failing. This fix iterates over the stopTimes for each pair of
trips until it finds one that exists for each pair and then it sorts on that index.

fixes catalogueglobal/datatools-server#48