feat(patterns): edit pattern geometry segment by segment
This update improves the pattern editing feature in the editor and adapts the UI to work with the
SQL backend. The backend represents the pattern geometry now as an array of shapepoints rather than
a simple GeoJSON linestring. This allows for saving metadata about the points, specifically whether
it is the projected point from a stop, a user-defined anchor point, or simple another coordinate in
the sequence. In the UI using selectors, we construct two arrays from these shape points: 1) an
array of control points to render for the user to manipulate and 2) an array of coordinate sets
(essentially a multi-linestring) frim which to render the geometry.

BREAKING CHANGE