feat(editor): restrict editing for a feed to a single session
This adds a procedure to "check out" a feed for editing. If the feed is already being edited (by
another user or in another tab), the user is prevented from editing it. When a feed is checked out
for editing, the UI will make calls to the server every ten seconds to indicate that the feed is
still checked out. When a user exits the Editor or closed the browser tab/window, a request is made
to the server to remove the lock on the feed.