fix(pattern): scale shape dist traveled when calc. pattern shape slices
The GTFS spec does not specify a unit for shape_dist_traveled. Previously, we were assuming these
values to be in meters. This fix allows original values to be in some other unit, scaling the
distance in meters values by a factor derived by dividing the pattern length in meters by the max
shape_dist_traveled value for the shape.

fixes #82