ci(e2e): use custom server for front end on travis
budo was watching way too many files and causing an error for doing so on Travis.  Files don't need to be watched on Travis but budo can't disable watching of files thus this server is needed.  Also, this approach is able to separate the build process from serving which enables the removal of arbitrary timeouts for the build process.