improvement(OatternStopCard.js lib/gtfs/util/index.js lib/types/index.js): Update to the GTFS Spec U
Added the changes where applicable to the gtfs.yml file. Added the new types for continuous drop
off/pick up and added fields as well as basic validation in the patternStopCard file.

BREAKING CHANGE: There have been added fields that need to coincide with changes on the backend as
the api will be looking for the new fields and those fields need to be present on the Front end

#663